trading-bot-mock-binance/requirements.txt
python-binance>=1.0.19
pytest>=7.0.0
